+++ packages/docusaurus-theme/src/theme/EditThisPage/index.tsx
@@ -5,7 +5,7 @@ import type { Props } from '@theme-original/EditThisPage';
 import { EuiButton, useEuiMemoizedStyles, UseEuiTheme } from '@elastic/eui';
 // @ts-ignore - eui only has module declarations for '@elastic/eui/src/themes/amsterdam/global_styling/mixins/button'
 // but importing from /src results in "Module not found" error
-import { euiButtonColor } from '@elastic/eui/lib/themes/amsterdam/global_styling/mixins/button';
+import { euiButtonColor } from '@elastic/eui/lib/global_styling/mixins/_button';
 
 const getStyles = (theme: UseEuiTheme) => {
   const { euiTheme } = theme;
