+++ packages/eui/src/components/button/button_empty/button_empty.tsx
@@ -26,7 +26,7 @@ import {
 import {
   useEuiButtonColorCSS,
   _EuiButtonColor,
-} from '../../../themes/amsterdam/global_styling/mixins/button';
+} from '../../../global_styling/mixins/_button';
 import { isButtonDisabled } from '../button_display/_button_display';
 
 import { euiButtonEmptyStyles } from './button_empty.styles';
