+++ packages/eui/src/components/button/button_group/button_group.tsx
@@ -18,7 +18,7 @@ import { useEuiMemoizedStyles } from '../../../services';
 import { EuiScreenReaderOnly } from '../../accessibility';
 import { CommonProps } from '../../common';
 
-import { _EuiButtonColor } from '../../../themes/amsterdam/global_styling/mixins';
+import { _EuiButtonColor } from '../../../global_styling/mixins';
 import { EuiToolTipProps } from '../../../components/tool_tip';
 import { EuiButtonDisplayContentProps } from '../button_display/_button_display_content';
 import { EuiButtonGroupButton } from './button_group_button';
