+++ packages/eui/src/components/button/button_icon/button_icon.tsx
@@ -30,7 +30,7 @@ import {
   useEuiButtonColorCSS,
   useEuiButtonFocusCSS,
   _EuiButtonColor,
-} from '../../../themes/amsterdam/global_styling/mixins/button';
+} from '../../../global_styling/mixins/_button';
 import { isButtonDisabled } from '../button_display/_button_display';
 import { euiButtonIconStyles, _emptyHoverStyles } from './button_icon.styles';
